feat(service):业务移到service层

+++ auth-server/src/main/java/com/tim/auth/service/impl/RoleServiceImpl.java
@@ -6,6 +6,7 @@ import com.tim.auth.dao.RoleMapper;
 import com.tim.auth.po.Role;
 import com.tim.auth.po.RoleExample;
 import com.tim.auth.po.RoleExample.Criteria;
+import com.tim.auth.sdk.constant.AuthConstant;
 import com.tim.auth.service.RoleMenuService;
 import com.tim.auth.service.RoleService;
 import com.tim.auth.service.RoleUserService;
@@ -17,9 +18,11 @@ import com.tim.auth.vo.RoleSearchResp;
 import com.tim.auth.vo.RoleUpdate;
 import com.tim.auth.vo.RoleUserAdd;
 import com.tim.auth.vo.RoleUserDel;
+import com.tim.message.Message;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang.StringUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -32,6 +35,7 @@ import org.springframework.transaction.annotation.Transactional;
  * @description：
  */
 @Service
+@Slf4j
 public class RoleServiceImpl implements RoleService {
 
   @Autowired
@@ -84,23 +88,33 @@ public class RoleServiceImpl implements RoleService {
   }
 
   @Override
-  public boolean add(RoleAdd roleAdd) {
+  public Message add(RoleAdd roleAdd) {
     Role role = new Role();
     BeanUtils.copyProperties(roleAdd, role);
     String id = UUID.randomUUID().toString();
     role.setId(id);
     role.setCreatorId(tokenManager.getUserId());
 
-    return roleMapper.insertSelective(role) == 1 ? true : false;
+    int row = roleMapper.insertSelective(role);
+    if (row == 1) {
+      return Message.success("新增角色成功！");
+    }
+
+    return Message.error("新增角色失败！");
   }
 
   @Override
-  public boolean update(RoleUpdate roleUpdate) {
+  public Message update(RoleUpdate roleUpdate) {
     Role role = new Role();
     BeanUtils.copyProperties(roleUpdate, role);
     role.setModifierId(tokenManager.getUserId());
 
-    return roleMapper.updateByPrimaryKeySelective(role) == 1 ? true : false;
+    int row = roleMapper.updateByPrimaryKeySelective(role);
+    if (row == 1) {
+      return Message.success("更新角色成功！");
+    }
+
+    return Message.error("更新角色失败！");
   }
 
   @Override
@@ -114,7 +128,12 @@ public class RoleServiceImpl implements RoleService {
 
   @Override
   @Transactional(rollbackFor = {RuntimeException.class, Error.class})
-  public boolean delete(String id) {
+  public Message delete(String id) {
+    if (id.equals(AuthConstant.ROLE_ADMIN_ID) || id.equals(AuthConstant.ROLE_COMMON_ID)) {
+      log.warn("系统内置角色不能删除，id:{}", id);
+      return Message.error("系统内置角色不能删除！");
+    }
+
     //从角色用户表，删除角色
     roleUserService.deleteRole(id);
 
@@ -127,34 +146,48 @@ public class RoleServiceImpl implements RoleService {
     //刷新redis
     loadResourceRole.load();
 
-    return true;
+    return Message.success("角色删除成功！");
   }
 
   @Override
-  public boolean addUser(RoleUserAdd roleUserAdd) {
+  public Message addUser(RoleUserAdd roleUserAdd) {
     //先删除旧的用户
     roleUserService.deleteRole(roleUserAdd.getRoleId());
 
     //刷新redis
     loadResourceRole.load();
+    roleUserService.addUser(roleUserAdd);
 
-    return roleUserService.addUser(roleUserAdd);
+    return Message.success();
   }
 
   @Override
-  public boolean addMenu(RoleMenuAdd roleMenuAdd) {
+  public Message addMenu(RoleMenuAdd roleMenuAdd) {
+    String roleId = roleMenuAdd.getRoleId();
+    if (roleId.equals(AuthConstant.ROLE_ADMIN_ID) || roleId.equals(AuthConstant.ROLE_COMMON_ID)) {
+      log.warn("系统内置角色权限不能修改，id:{}", roleId);
+      return Message.error("系统内置角色权限不能修改！");
+    }
+
     //先删除旧的菜单分配
     roleMenuService.deleteRole(roleMenuAdd.getRoleId());
 
     //刷新redis
     loadResourceRole.load();
 
-    return roleMenuService.addMenu(roleMenuAdd);
+    roleMenuService.addMenu(roleMenuAdd);
+
+    return Message.success();
   }
 
   @Override
-  public boolean deleteUser(RoleUserDel roleUserDel) {
-    return roleUserService.deleteUser(roleUserDel);
+  public Message deleteUser(RoleUserDel roleUserDel) {
+    boolean result = roleUserService.deleteUser(roleUserDel);
+    if (!result) {
+      return Message.error();
+    }
+
+    return Message.success();
   }
 
   @Override
